requirements.txt
azure-storage-blob==12.4.0
Django==2.2
requests==2.24.0
pyswagger
django-crontab==0.7.1
pandas==1.2.4
django-auth-adfs==1.9.0
./python-client